SB_Galleon/player/625: keep only the cast sites that actually block a CSE
A redundant same-type cast is a value-numbering barrier, not a copy: it is a
machine no-op when a base is spelled uniformly, and only bites where the same
lvalue is read twice with the two different spellings. Per-site grading finds
38 of the 40 SB_Galleon_updateFlight object casts inert; the two that stay sit
against plain-spelled reads of the same members.

+++ src/dlls/objects/625/625.c
@@ -913,7 +913,7 @@ void drakorhoverpad_updateMain(GameObject* obj)
             {
                 yawDelta = 0x200;
             }
-            c = (s16)yawDelta;
+            c = yawDelta;
             (obj)->anim.rotX += (s16)c;
             if ((obj)->anim.rotY != 0)
             {
